fix delete topic

+++ src/schema/schema.ts
@@ -17,33 +17,33 @@ const Query = new GraphQLObjectType({
     word: {
       type: WordType,
       args: { id: { type: GraphQLID } },
-      resolve(parent, args) {
+      resolve(_parent, args) {
         return Word.findById(args.id);
       },
     },
     words: {
       type: new GraphQLList(WordType),
-      resolve(parent, args) {
+      resolve(_parent, args) {
         return Word.find({});
       },
     },
     wordsByTopicId: {
       type: new GraphQLList(WordType),
       args: { topicId: { type: GraphQLID } },
-      resolve(parent, { topicId }) {
+      resolve(_parent, { topicId }) {
         return Word.find({ topicId });
       },
     },
     topic: {
       type: TopicType,
       args: { id: { type: GraphQLID } },
-      resolve(parent, args) {
+      resolve(_parent, args) {
         return Topic.findById(args.id);
       },
     },
     topics: {
       type: new GraphQLList(TopicType),
-      resolve(parent, args) {
+      resolve(_parent, args) {
         return Topic.find({});
       },
     },
@@ -61,7 +61,7 @@ const Mutation = new GraphQLObjectType({
         example: { type: GraphQLString },
         topicId: { type: GraphQLID },
       },
-      resolve(parent, { word, translate, example, topicId }) {
+      resolve(_parent, { word, translate, example, topicId }) {
         const newWord = new Word({ word, translate, example, topicId });
         return newWord.save();
       },
@@ -71,7 +71,7 @@ const Mutation = new GraphQLObjectType({
       args: {
         name: { type: GraphQLString },
       },
-      resolve(parent, { name }) {
+      resolve(_parent, { name }) {
         const newTopic = new Topic({ name });
         return newTopic.save();
       },
@@ -79,15 +79,21 @@ const Mutation = new GraphQLObjectType({
     deleteWord: {
       type: WordType,
       args: { id: { type: GraphQLID } },
-      resolve(parent, { id }) {
+      resolve(_parent, { id }) {
         return Word.findByIdAndRemove(id);
       },
     },
     deleteTopic: {
       type: TopicType,
       args: { id: { type: GraphQLID } },
-      resolve(parent, { id }) {
-        return Topic.findByIdAndRemove(id);
+      resolve(_parent, { id }) {
+        const someWord = Word.findOne({ topicId: id });
+
+        if (someWord) {
+          throw new Error("Words with current topicId exist");
+        } else {
+          return Topic.findByIdAndRemove(id);
+        }
       },
     },
     updateWord: {
@@ -99,7 +105,7 @@ const Mutation = new GraphQLObjectType({
         example: { type: GraphQLString },
         topicId: { type: GraphQLString },
       },
-      resolve(parent, { id, word, translate, example, topicId }) {
+      resolve(_parent, { id, word, translate, example, topicId }) {
         return Word.findByIdAndUpdate(id, {
           word,
           translate,
@@ -114,7 +120,7 @@ const Mutation = new GraphQLObjectType({
         id: { type: GraphQLID },
         name: { type: GraphQLString },
       },
-      resolve(parent, { id, name }) {
+      resolve(_parent, { id, name }) {
         return Topic.findByIdAndUpdate(id, { name });
       },
     },
